% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suave_function.R
\name{Suave_function}
\alias{Suave_function}
\title{A Suaveometer Function}
\usage{
Suave_function(df)
}
\description{
This function allows you to know how Suave you are on Tinder.
Suave_function()
}

